Adding goofy namespace
namespace is called LostResort.Passengers
