--- GMTK 2025/Assets/PassengerScripts/ResortLocation.cs
@@ -1,18 +1,0 @@
-using UnityEngine;
-
-public class ResortLocation : MonoBehaviour
-{
-   [SerializeField] private LocationType locationType;
-   
-   public LocationType GetLocationType()
-    {
-        return locationType;
-    }
-
-}
-public enum LocationType{
-   Lobby,
-   Villas,
-   Beach,
-   Restaurants
-}
