+++ GMTK 2025/Assets/Scripts/Passengers/LostResort.Passengers.asmdef
@@ -1,0 +1,3 @@
+{
+	"name": "LostResort.Passengers"
+}
