+++ GMTK 2025/Assets/Scripts/Passengers/LostResort.Passengers.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b5669313068bde54e9aed42bffee183e
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
