+++ GMTK 2025/Assets/Scripts/Passengers/TemporaryColorMaterials/Blue.mat
