+++ GMTK 2025/Assets/Scripts/Passengers/TemporaryColorMaterials/Red.mat
