+++ GMTK 2025/Assets/Scripts/Passengers/TemporaryColorMaterials/Yellow.mat.meta
